+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Change Log
 11. Added window.app as an alias to window.Application. (XDK compatible)
 12. Added release/edge build tag into app.stagejs.
 13. Fixed error in ro type editor enable/disable ops.
+14. Fixed error in form view (item-view) setValues (allow 0, '' & false).
+15. Added cordova watch to devserver watchers. (Unix-like system only)
 
 
 In progress...